Remove drop down
