+++ client/components/users/UserProfile.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router';
 import ListOfEvents from './ListOfEvents.js';
-import JoinAnEvent from './JoinAnEvent.js';
 
 require('./../../styles/styles.css');
 
@@ -11,25 +10,11 @@ export default class UserProfile extends React.Component {
     super(props);
   }
 
-  joinEvent() {
-    console.log('hi');
-    $.ajax({
-      type: 'GET',
-      url: '/api/getEvents',
-      dataType: 'json',
-      success: function(data) {
-        console.log('Data in joinEvent', data)
-      }.bind(this),
-      fail: function(err) {
-        console.error(err);
-      }
-    });
-  }
-
   render() {
     const {
       username,
-      events
+      events,
+      allEvents
     } = this.props
     return (
       <div className="profile">
@@ -38,8 +23,6 @@ export default class UserProfile extends React.Component {
         <div className="logout-btn">
           <a href='/logout' className="waves-effect waves-light btn style-btn"><span className='btn-text'>Logout</span></a>
         </div>
-        {/*Join an event*/}
-        <JoinAnEvent joinEvent={this.joinEvent.bind(this)} events={events} />
         {/*Create an event button*/}
         <div className="add-btn btn-floating btn-large waves-effect waves-light blue"><Link to='/events'><i className="material-icons">add</i></Link></div>
         {/*<Link to='/events/abc'>Test</Link>*/}
